+++ NEWS.rst
@@ -24,7 +24,7 @@ Version 0.7.0 (upcoming)
 
   * Ludl MC 2000 (:class:`microscope.controllers.ludl.LudlMC2000`)
 
-  * Raspberry Pi picam (:class:`microscope.cameras.picam.Picamera`)
+  * Raspberry Pi picam (:class:`microscope.cameras.picam.PiCamera`)
 
   * Toptica iChrome MLE (:class:`microscope.controllers.toptica.iChromeMLE`)